ref #12108 retry in bgp socket connect
*currently bgp schedules an event to zebra socket connect.
if bgp config is processed before the socket establishment,
then all the neighbor registeration will stall in bgp, as bgp
doesn't handle the rnh register failure with zebra.

*Added the zebra socket retry in bgp to retry for 12 times
with 10 seconds interval.

*This issue observed in a cluster where the storage is being
put in a cross regional account. it could cause the issue
as we aread the inf_config.json file in the zebra bootup.

*adding the patch to dev folder for future reference.