Draft Initial changes for BWAN-2309 overlay suboptimal routing
Need click side changes to generate overlay metric considering
end to end path reachability
Currently metric is inferred from overlay and nh counters which works
to prefer optimal hub. For spoke to spoke cases we need click side
changes to generate the overlay metric.
Code is disabled as of now with a flag